Decoupled feed service into service and utils.
